Update The City
The City

The sun rises over the city buildings, casting a bright light over the landscape. Uncovering the vast activities available for entertainment. Sketched artwork as far as the eyes can see. 

The city, a place of raw beauty and danger comes alive with the sounds of vibrations. 

This is a land of endless potential. The managers stand tall and proud. 

The citizens provide shade and shelter to the people that inherit this vast landscape. 

Like precious gems, a lifeline in this environment of endless potential. Drawing people from miles, a testament to the cycle of life and death that plays out on this ancient stage.

The vibrations brakes through the walls carrying with it's a scent. A whispers of stories of survival and endurance. 

Every step holds the memory of the countless generations that have thrive in parish. 

The facility is a place where survival is a daily struggle, and only the fittest survive. Managers and customers engage in internal dance. Each playing their part in the delicate balance of this ecosystem. The hunt is a spectacle of raw power and instinct. A reminder of the realities of life in the facility.

As the day progresses, the heat of the sun intensifies and the air is filled with a distant roar. Each sound tells a story in the saga of civilization. 

The facility, with it’s essential managers, are the undisputed kings of their realms. 

The employees are a close knit family unit. Working together to hunt and protect their territory. Their services across the city is a powerful reminder of their dominance.